compile.c: suppress a warning
* compile.c (iseq_build_kw): suppress an alloc-size-larger-than
  warning by gcc 7.

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@57617 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ compile.c
@@ -6892,7 +6892,7 @@ iseq_build_kw(rb_iseq_t *iseq, VALUE params, VALUE keywords)
 	return keyword;
     }
 
-    dvs = ALLOC_N(VALUE, default_len);
+    dvs = ALLOC_N(VALUE, (unsigned int)default_len);
 
     for (j = 0; i < len; i++, j++) {
 	key = RARRAY_AREF(keywords, i);
